requirements.txt
requests
rich
Pillow
